requirements.txt
arrow==0.12.1
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.6.3
binaryornot==0.4.4
bs4==0.0.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cli-helpers==1.1.0
Click==7.0
configobj==5.0.6
constantly==15.1.0
cookiecutter==1.6.0
cryptography==2.3
cssselect==1.0.3
decorator==4.3.0
future==0.17.1
humanize==0.5.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
Jinja2==2.10
jinja2-time==0.2.0
lxml==4.2.4
MarkupSafe==1.1.0
parsel==1.5.0
parso==0.3.1
pexpect==4.6.0
pgcli==2.0.0
pgspecial==1.11.2
pickleshare==0.7.5
pip-autoremove==0.9.1
poyo==0.4.2
prompt-toolkit==2.0.5
psycopg2==2.7.5
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-dateutil==2.7.5
queuelib==1.5.0
requests==2.20.0
Scrapy==1.5.1
scrapy-splash==0.7.2
scrapyd==1.2.0
scrapyd-client==1.2.0a1
service-identity==17.0.0
setproctitle==1.1.10
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.12
sqlparse==0.2.4
tabulate==0.8.2
terminaltables==3.1.0
traitlets==4.3.2
Twisted==18.7.0
urllib3==1.24.1
w3lib==1.19.0
wcwidth==0.1.7
whichcraft==0.5.2
zope.interface==4.5.0
